+++ app/Models/Article.php
@@ -20,6 +20,7 @@ use Spatie\Sluggable\SlugOptions;
  * @property int $category_id
  * @property \Illuminate\Support\Carbon|null $created_at
  * @property \Illuminate\Support\Carbon|null $updated_at
+ * @property-read \App\Models\Category $category
  * @property-read string $image
  * @property-read string $status_class
  * @property-read string $status_name
@@ -45,7 +46,7 @@ class Article extends Model
     /**
      * The attributes that are mass assignable.
      *
-     * @var array
+     * @var array<int, string>
      */
     protected $fillable = [
         'name',