+++ resources/js/bootstrap.js
@@ -30,6 +30,9 @@ window.Flasher = require("@flasher/flasher");
 import Chart from "chart.js/auto";
 window.Chart = Chart;
 
+import { PieChart } from "./charts/PieChart";
+window.PieChart = PieChart;
+
 /**
  * Echo exposes an expressive API for subscribing to channels and listening
  * for events that are broadcast by Laravel. Echo and event broadcasting