+++ resources/views/livewire/series/volumes-table.blade.php
@@ -55,7 +55,7 @@
                 @endforeach
                 @if ($volumes->count() == 0)
                     <tr>
-                        <td colspan="5" style="text-align: center;">{{ __('No data') }}</td>
+                        <td colspan="@if ($enable_reordering) 9 @else 8 @endif" style="text-align: center;">{{ __('No data') }}</td>
                     </tr>
                 @endif
             </tbody>